pyproject.toml
[project]
name = "jarvis-assistant"
version = "0.1.0"
description = "A Linux assistant that lives in your terminal."
authors = [
    {name = "garvit", email = "garvitjoshi543@gmail.com"},
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "distro>=1.8.0",
    "google-genai>=0.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "browser-use>=0.1.0",
    "browser-use-sdk>=0.0.1",
    "playwright>=1.40.0",
    "langchain-google-genai>=0.0.1",
    "langchain-openai>=0.0.1"
]
requires-python = ">=3.10"

[project.scripts]
nexus = "jarvis.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]
